app_api/requirements.txt
alembic==1.8.1
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
ecdsa==0.18.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
prompt-toolkit==3.0.33
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.6
redis==4.3.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.5